fix: move drag indicator reducer out of component file
Satisfies react-refresh/only-export-components by colocating
reduceControlMatDragIndicator in controlMatDragIndicatorState.ts.
