+++ src/shared/ui/NotebookShell/index.ts
@@ -2,13 +2,15 @@ export {
   ControlMat
 } from './chrome/ControlMat';
 export {
-  ControlMatDragIndicator,
-  reduceControlMatDragIndicator
+  ControlMatDragIndicator
 } from './chrome/ControlMatDragIndicator';
+export {
+  reduceControlMatDragIndicator
+} from './chrome/controlMatDragIndicatorState';
 export type {
   ControlMatDragIndicatorAction,
   ControlMatDragIndicatorState
-} from './chrome/ControlMatDragIndicator';
+} from './chrome/controlMatDragIndicatorState';
 export {
   NotebookHeaderChrome
 } from './chrome/NotebookHeaderChrome';
